% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{doepro_server}
\alias{doepro_server}
\title{The DOEpro server logic}
\usage{
doepro_server(input, output, session)
}
\arguments{
\item{input, output, session}{Standard Shiny server arguments.}
}
\value{
Invisibly \code{NULL}; called for its side effects.
}
\description{
The Shiny server function. Called by \code{\link{run_DOEpro}}; you should not
normally need to call it yourself.
}
\keyword{internal}
