% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{doepro_ui}
\alias{doepro_ui}
\title{The DOEpro user interface}
\usage{
doepro_ui()
}
\value{
A Shiny UI definition.
}
\description{
Builds the Shiny UI object. Called by \code{\link{run_DOEpro}}; you should not
normally need to call it yourself.
}
\keyword{internal}
